pyproject.toml
[project]
name = "clyre"
version = "0.0.1-2"
description = "Pre-Alpha-2"
authors = [
    { name = "Yevhenii Prasol", email = "mefgner@gmail.com" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "fastapi (>=0.116.2,<0.117.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pooch (>=1.8.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "colorlog (>=6.10.1,<7.0.0)"
]

#[build-system]
#requires = ["poetry-core>=2.0.0,<3.0.0"]
#build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "black (>=25.1.0,<26.0.0)",
    "pylint (>=3.3.8,<4.0.0)",
    "ruff (>=0.13.0,<0.14.0)"
]

[tool.black]
line-length = 96
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | build
  | dist
  | web
  | node_modules
  | telegram-bot
  | public
  | configs
)/
'''

[tool.ruff]
line-length = 96
target-version = "py311"
src = ["api"]
exclude = [
    "web",
    "node_modules",
    "telegram-bot",
    "public",
    "configs",
    ".venv",
    "dist",
    "build",
]
